app/requirements.txt
alembic==1.0.0
appnope==0.1.0
Babel==2.6.0
backcall==0.1.0
certifi==2018.8.13
chardet==3.0.4
click==6.7
decorator==4.3.0
dominate==2.3.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.10
traitlets==4.3.2
urllib3==1.23
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
